test(categry): finished switch test module category

+++ app/Livewire/Panel/Category/Delete.php
@@ -24,6 +24,10 @@ class Delete extends Component
     }
     public function confirmed(string $message): void
     {
+        if (empty($this->category)) {
+
+            throw 'Error in delete category';
+        }
         $this->category->delete();
         $this->dispatch('category:deleted');
         $this->dialog()->success('Sucesso', $message)->send();
